Update wallet.sh
After making the first transaction on the Seismic devnet, finding the wallet private key required a slow process of opening the wallet.sh file and extracting it from there.

I added a function to display the wallet address, and the private key is now shown in the bottom column next to the address. This makes it easier for newcomers to interact with the ecosystem.